+++ src/client/pages/user/index.photos.vue
@@ -2,11 +2,11 @@
 <div class="ujigsodd">
 	<mk-loading v-if="fetching"/>
 	<div class="stream" v-if="!fetching && images.length > 0">
-		<a v-for="(image, i) in images" :key="i"
+		<router-link v-for="(image, i) in images" :key="i"
 			class="img"
 			:style="`background-image: url(${thumbnail(image.file)})`"
-			:href="image.note | notePage"
-		></a>
+			:to="image.note | notePage"
+		></router-link>
 	</div>
 	<p class="empty" v-if="!fetching && images.length == 0">{{ $t('nothing') }}</p>
 </div>