+++ src/remote/activitypub/type.ts
@@ -67,6 +67,15 @@ export interface IActivity extends IObject {
 	actor: IObject | string;
 	object: IObject | string;
 	target?: IObject | string;
+	/** LD-Signature */
+	signature?: {
+		type: string;
+		created: Date;
+		creator: string;
+		domain?: string;
+		nonce?: string;
+		signatureValue: string;
+	};
 }
 
 export interface ICollection extends IObject {
@@ -111,10 +120,10 @@ interface IQuestionChoice {
 	_misskey_votes?: number;
 }
 
-export const validActor = ['Person', 'Service', 'Group', 'Organization'];
+export const validActor = ['Person', 'Service', 'Group', 'Organization', 'Application'];
 
 export interface IPerson extends IObject {
-	type: 'Person';
+	type: 'Person' | 'Service' | 'Organization' | 'Group' | 'Application';
 	name?: string;
 	preferredUsername?: string;
 	manuallyApprovesFollowers?: boolean;
@@ -234,6 +243,10 @@ export interface IBlock extends IActivity {
 	type: 'Block';
 }
 
+export interface IFlag extends IActivity {
+	type: 'Flag';
+}
+
 export const isCreate = (object: IObject): object is ICreate => object.type === 'Create';
 export const isDelete = (object: IObject): object is IDelete => object.type === 'Delete';
 export const isUpdate = (object: IObject): object is IUpdate => object.type === 'Update';
@@ -247,3 +260,4 @@ export const isRemove = (object: IObject): object is IRemove => object.type === 
 export const isLike = (object: IObject): object is ILike => object.type === 'Like' || object.type === 'EmojiReaction' || object.type === 'EmojiReact';
 export const isAnnounce = (object: IObject): object is IAnnounce => object.type === 'Announce';
 export const isBlock = (object: IObject): object is IBlock => object.type === 'Block';
+export const isFlag = (object: IObject): object is IFlag => object.type === 'Flag';